+++ lib/src/player.dart
@@ -80,7 +80,7 @@ class Player {
       return true;
     }
     _respawnCount -= 1;
-    print(_respawnCount);
+    //print(_respawnCount);
     return false;
   }
 }